requirements.txt
python3.10.15
torch2.1.2+cu118
torchaudio2.1.2+cu118
torchvision0.16.2+cu118
numpy1.26.3
scipy1.15.1
torch-geometric2.6.1
ipdb0.13.13
pypower5.1.17
pyrlu0.2.1
GitPython3.1.44
